more develop.

+++ proxy-core/src/main/java/m2/proxy/tcp/handlers/ConnectionWorkCount.java
@@ -1,6 +1,7 @@
 package m2.proxy.tcp.handlers;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class ConnectionWorkCount {
 
@@ -9,7 +10,7 @@ public class ConnectionWorkCount {
     public AtomicInteger message = new AtomicInteger();
     public AtomicInteger request = new AtomicInteger();
     public AtomicInteger reply = new AtomicInteger();
-    public AtomicInteger bytes = new AtomicInteger();
+    public AtomicLong bytes = new AtomicLong();
 
 
 }
